requirements.txt
annotated-types==0.7.0
anyio==4.8.0
bcrypt==4.1.2
certifi==2025.1.31
cffi==1.17.1
click==8.1.8
distro==1.9.0
dnspython==2.7.0
ecdsa==0.19.1
email-validator==2.1.0
exceptiongroup==1.2.2
fastapi==0.109.0
greenlet==3.1.1
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.10
iniconfig==2.0.0
jiter==0.9.0
joblib==1.4.2
lxml==5.3.1
numpy==2.2.3
openai==1.66.3
oso==0.27.3
packaging==24.2
passlib==1.7.4
pluggy==1.5.0
pyasn1==0.6.1
pycparser==2.22
pydantic==2.5.3
pydantic-settings==2.1.0
pydantic_core==2.14.6
pypdf==5.4.0
pytest==8.3.5
pytest-asyncio==0.25.3
python-docx==1.1.2
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.6
rsa==4.9
scikit-learn==1.6.1
scipy==1.15.2
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.25
starlette==0.35.1
threadpoolctl==3.6.0
tomli==2.2.1
tqdm==4.67.1
typing_extensions==4.12.2
uvicorn==0.27.0
